requirements.txt
streamlit
openai
google-genai
langchain
langchain-openai
langchain-google-genai